Rename form_avatar_controller in form_file_controller
