+++ src/assets/javascripts/application.js
@@ -7,7 +7,7 @@ import BackStorageController from 'js/controllers/back_storage_controller.js';
 import CollectionsSelectorController from 'js/controllers/collections_selector_controller.js';
 import ConfirmationController from 'js/controllers/confirmation_controller.js';
 import CopyToClipboardController from 'js/controllers/copy_to_clipboard_controller.js';
-import FormAvatarController from 'js/controllers/form_avatar_controller.js';
+import FormFileController from 'js/controllers/form_file_controller.js';
 import InputPasswordController from 'js/controllers/input_password_controller.js';
 import ModalController from 'js/controllers/modal_controller.js';
 import ModalOpenerController from 'js/controllers/modal_opener_controller.js';
@@ -27,7 +27,7 @@ application.register('back-storage', BackStorageController);
 application.register('collections-selector', CollectionsSelectorController);
 application.register('confirmation', ConfirmationController);
 application.register('copy-to-clipboard', CopyToClipboardController);
-application.register('form-avatar', FormAvatarController);
+application.register('form-file', FormFileController);
 application.register('input-password', InputPasswordController);
 application.register('modal', ModalController);
 application.register('modal-opener', ModalOpenerController);
